+++ src/core/utils/html.js
@@ -5,3 +5,13 @@ export function escapeHtml(str) {
         .replace(/</g, '&lt;')
         .replace(/>/g, '&gt;');
 }
+
+export function numDelim(num) {
+    const strNum = String(num);
+
+    if (strNum.length > 3) {
+        return strNum.replace(/\..+$|\B(?=(\d{3})+(\D|$))/g, m => m || '<span class="num-delim"></span>');
+    }
+
+    return strNum;
+}