+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `destroy()` method to Popup view
 - Added `external` data option for `badge` views
 - Fixed a crush on open actions popup in `struct` view when a value has too much nested structure
+- Uniformed a monospace font family across all views
 
 ## 1.0.0-beta.65 (18-04-2022)